MicroServices-based-DeepFake-Detection-System-using-Docker/video_service/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
transformers==4.39.3
torch==2.3.1+cpu
torchvision==0.18.1+cpu
numpy>=1.21.0
opencv-python-headless>=4.5.0
Pillow>=9.0.0
python-multipart>=0.0.5
--extra-index-url https://download.pytorch.org/whl/cpu